Update gatsby and react
Update some dependencies, including gatsby, gatsby plugins, as well as react.

In the future, it would be advisable to avoid relying on version tags like "next" which are subject to change at any time by the project, which is what happened here. The "next" npm dist-tag was removed breaking the package.json file.